+++ AWSCore/Service/AWSServiceEnum.h
@@ -91,6 +91,10 @@ typedef NS_ENUM(NSInteger, AWSRegionType) {
      *  AWS GovCloud (US)
      */
     AWSRegionUSGovWest1 NS_SWIFT_NAME(USGovWest1),
+    /**
+     *  China (Ningxia)
+     */
+    AWSRegionCNNorthWest1 NS_SWIFT_NAME(CNNorthWest1),
 };
 
 /**