+++ AWSCore/Utility/AWSCategory.m
@@ -565,6 +565,12 @@ static NSTimeInterval _clockskew = 0.0;
         || [self isEqualToString:@"cn-north-1"]) {
         return AWSRegionCNNorth1;
     }
+    
+    if ([self isEqualToString:@"AWSRegionCNNorthWest1"]
+        || [self isEqualToString:@"CNNorthWest1"]
+        || [self isEqualToString:@"cn-northwest-1"]) {
+        return AWSRegionCNNorthWest1;
+    }
 
     return AWSRegionUnknown;
 }